Moved system info tests outside test_common
