+++ tests/frame/test_common.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import warnings
 from typing import Any
 
 import pandas as pd
@@ -9,9 +8,6 @@ import pyarrow as pa
 import pytest
 
 import narwhals.stable.v1 as nw
-from narwhals.functions import _get_deps_info
-from narwhals.functions import _get_sys_info
-from narwhals.functions import show_versions
 from tests.utils import compare_dicts
 
 data = {"a": [1, 3, 2], "b": [4, 4, 6], "z": [7.0, 8, 9]}
@@ -136,41 +132,3 @@ def test_with_columns_order_single_row(constructor_eager: Any) -> None:
     assert result.columns == ["a", "b", "z", "d"]
     expected = {"a": [2], "b": [4], "z": [7.0], "d": [0]}
     compare_dicts(result, expected)
-
-
-def test_get_sys_info() -> None:
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore")
-        show_versions()
-        sys_info = _get_sys_info()
-
-    assert "python" in sys_info
-    assert "executable" in sys_info
-    assert "machine" in sys_info
-
-
-def test_get_deps_info() -> None:
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore")
-        show_versions()
-        deps_info = _get_deps_info()
-
-    assert "narwhals" in deps_info
-    assert "pandas" in deps_info
-    assert "polars" in deps_info
-    assert "cudf" in deps_info
-    assert "modin" in deps_info
-    assert "pyarrow" in deps_info
-    assert "numpy" in deps_info
-
-
-def test_show_versions(capsys: Any) -> None:
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore")
-        show_versions()
-        out, _ = capsys.readouterr()
-
-    assert "python" in out
-    assert "machine" in out
-    assert "pandas" in out
-    assert "polars" in out
